fix: speed up form list
